Revert increases in stim param limits
Empirical testing and subsequent examination of the code and documentation showed that hardware limitations prevent the total stim pulse duration from exceeding 65535 sample periods, i.e., about 2 or 3 seconds depending on the sampling rate. The number of pulse is also greatly limited by either the hardware or the Opal Kelly API; that parameter was already permitted to reach its maximum possible value of 256.